框架搭好了，可以注册和登陆

+++ src/main/java/com/hz/dafeiji/ai/user/modules/player/UserBaseInfo.java
@@ -6,12 +6,14 @@ import lombok.Data;
 /**
  * user         LIUKUN
  * time         2014-12-19 15:10
+ * 如果数据库没有对应的值，那么查询的结果就是缺省值
+ * 数字一般为0，字符串为null！！！！！
  */
 @Data
 public class UserBaseInfo{
-    private String                  userName;
-    private int                     uid;
-    private String                  nickName;
-    private UserStatus              userStatus;
-    private String                  pass;
+    private String userName;
+    private long uid;
+    private String nickName;
+    private UserStatus userStatus;
+    private String pass;
 }
